requirements.txt
alembic==1.11.1
aniso8601==9.0.1
black==23.3.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
